moved lite_forms/generators to core/views
