+++ exporter/applications/forms/common.py
@@ -17,7 +17,7 @@ from lite_forms.components import (
     Label,
     Checkboxes,
 )
-from lite_forms.generators import success_page
+from core.views import success_page
 from lite_forms.helpers import conditional
